client side pagination with react tables
